Re-organizing code from the store and starting logic for guess submission
