+++ game/api/index.ts
@@ -1,1 +1,2 @@
+export { getResult } from "./getResult";
 export { sampleSize } from "./sampleSize";
