+++ game/constants/game.ts
@@ -1,1 +1,2 @@
 export const WORD_LENGTH = 5;
+export const NUMBER_OF_TRIES = 5;
